requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==5.0.1
Django==4.2.7
django-allauth==0.58.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
multidict==6.0.4
oauthlib==3.2.2
packaging==23.2
phonenumbers==8.13.24
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
twilio==8.10.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
whitenoise==6.6.0
yarl==1.9.2
